+++ unclassifiedFlower.cpp
@@ -1,2 +1,2 @@
 #include "unclassifiedFlower.hpp"
-UnclassifiedFlower::UnclassifiedFlower(int x1,int x2,int x3,int x4):Flower(x1,x2,x3,x4){}
+UnclassifiedFlower::UnclassifiedFlower(double x1,double x2,double x3,double x4):Flower(x1,x2,x3,x4){}
